chg: Rename wellcomeml.spacy to wellcomeml.prodigy
