--- wellcomeml/spacy/__init__.py
@@ -1,1 +1,0 @@
-from .spacy_doc_to_prodigy import SpacyDocToProdigy
